Move ChangePasswordDto
